more scripting features and multithreading support
added multithreaded support for scripts and messaging components, also added a few scripting features like editting and adding named walls to the map and launching scripts. added a few core message types like echoMessage, that resends the contents of the message as individual messages and triggerNotification, which triggers an out of schedule notification to compensate for the interrumpted nature of the new messaging system.